requirements.txt
chex==0.1.82
distrax==0.1.4
dm-haiku==0.0.5
entmax-jax @ git+https://github.com/deep-spin/entmax-jax@55b4644
etils==1.3.0
flax==0.6.10
jraph==0.0.6.dev0
jumanji==0.3.1
learned-optimization @ git+https://github.com/google/learned_optimization.git@463ab9a
matplotlib==3.7.1
mlflow==2.3.2
optax==0.1.5
orbax-checkpoint==0.3.5
rlax==0.1.6
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.1+cpu
torch-geometric==2.4.0
tqdm
